Firebase Stub refactoring
